set up jwt auth

+++ backend/controllers/clients.js
@@ -4,6 +4,7 @@
 
 /* require modules
 ---------------------------------------------------------- */
+const jwt = require('jwt-simple')
 const express = require('express')
 const router = express.Router()
 
@@ -11,28 +12,82 @@ const router = express.Router()
 ---------------------------------------------------------- */
 const db = require('../models')
 
+/* require jwt config
+--------------------------------------------------------------- */
+const config = require('../../jwt.config.js')
+
+/* jwt middleware
+--------------------------------------------------------------- */
+const authMiddleware = (req, res, next) => {
+    const token = req.headers.authorization;
+    if (token) {
+        try {
+            const decodedToken = jwt.decode(token, config.jwtSecret);
+            req.user = decodedToken;
+            next();
+        } catch (err) {
+            res.status(401).json({ message: 'Invalid token' });
+        }
+    } else {
+        res.status(401).json({ message: 'Missing or invalid Authorization header' });
+    }
+};
+
 /* routes
 ---------------------------------------------------------- */
-// create client (client & admin access only)
-router.post('/', (req, res) => {
+// create client (signup route)
+router.post('/signup', (req, res) => {
     db.Client.create(req.body)
-        .then(client => res.json(client))
+        .then(client => {
+            const token = jwt.encode({ id: client.id }, config.jwtSecret)
+            res.json({ token: token })
+        })
+        .catch(() => {
+            res.status(401)
+                .json({ message: 'Could not create a new musician. Your email may have been already taken, or your password could be too weak. Try again.' })
+        })
+})
+
+// login route
+router.post('/login', async (req, res) => {
+    const foundClient = await db.Client.findOne({ email: req.body.email })
+    if (foundClient && foundClient.password === req.body.password) {
+        const payload = { id: foundClient.id }
+        const token = jwt.encode(payload, config.jwtSecret)
+        res.json({
+            token: token,
+            email: foundClient.email
+        })
+    } else {
+        res.status(401)
+	    .json({ message: 'Could not find musician with that email/password' })
+    }
 })
 
 // edit client (client & admin access only)
-router.put('/:clientId', (req, res) => {
-    db.Client.findByIdAndUpdate(
-        req.params.clientId,
-        req.body,
-        { new: true }
-    )
-        .then(client => res.json(client))
+router.put('/:clientId', authMiddleware, async (req, res) => {
+    const currentClient = await db.Client.findById(req.params.clientId)
+    if (currentClient.clientId == req.user.id) {
+        const newClient = await db.Client.findByIdAndUpdate(
+            req.params.clientId,
+            req.body,
+            { new: true }
+        )
+        res.json(newClient)
+    } else {
+        res.status(401).json({ message: 'Invalid user or token' })
+    }
 })
 
-// delete request (client & admin access only)
-router.delete('/:clientId', (req, res) => {
-    db.Client.findByIdAndDelete(req.params.clientId)
-        .then(() => res.json({ deletedClientId: req.params.clientId }))
+// delete client (client & admin access only)
+router.delete('/:clientId', authMiddleware, async (req, res) => {
+    const currentClient = await db.Client.findById(req.params.clientId)
+    if (currentClient.clientId == req.user.id) {
+        const deletedClient = await db.Client.findByIdAndDelete(req.params.clientId)
+        res.send('You deleted client ' + deletedClient._id)
+    } else {
+        res.status(401).json({ message: 'Invalid user or token' });
+    }
 })
 
 /* export to server.js
