+++ backend/controllers/reviews.js
@@ -4,12 +4,38 @@
 
 /* require modules
 ---------------------------------------------------------- */
+const jwt = require('jwt-simple')
 const express = require('express')
 const router = express.Router()
 
 /* db connection, models
 ---------------------------------------------------------- */
 const db = require('../models')
+
+/* require jwt config
+--------------------------------------------------------------- */
+const config = require('../../jwt.config.js')
+
+/* require jwt config
+--------------------------------------------------------------- */
+const config = require('../../jwt.config.js')
+
+/* jwt middleware
+--------------------------------------------------------------- */
+const authMiddleware = (req, res, next) => {
+    const token = req.headers.authorization;
+    if (token) {
+        try {
+            const decodedToken = jwt.decode(token, config.jwtSecret);
+            req.user = decodedToken;
+            next();
+        } catch (err) {
+            res.status(401).json({ message: 'Invalid token' });
+        }
+    } else {
+        res.status(401).json({ message: 'Missing or invalid Authorization header' });
+    }
+};
 
 /* routes
 ---------------------------------------------------------- */
@@ -26,25 +52,38 @@ router.get('/:musicianId', function (req, res) {
 })
 
 // create review (client & admin access only)
-router.post('/', (req, res) => {
-    db.Review.create(req.body)
+router.post('/', authMiddleware, (req, res) => {
+    db.Review.create({
+        ...req.body,
+        senderId: req.user.id
+    })
         .then(review => res.json(review))
 })
 
 // edit review (client & admin access only)
-router.put('/:reviewId', (req, res) => {
-    db.Review.findByIdAndUpdate(
-        req.params.reviewId,
-        req.body,
-        { new: true }
-    )
-        .then(review => res.json(review))
+router.put('/:reviewId', authMiddleware, async (req, res) => {
+    const userReview = await db.Review.findById(req.params.reviewId)
+    if (userReview.senderId == req.user.id) {
+        const newReview = await db.Review.findByIdAndUpdate(
+            req.params.reviewId,
+            req.body,
+            { new: true }
+        )
+        res.json(newReview)
+    } else {
+        res.status(401).json({ message: 'Invalid user or token' })
+    }
 })
 
 // delete review (client & admin access only)
-router.delete('/:reviewId', (req, res) => {
-    db.Review.findByIdAndDelete(req.params.reviewId)
-        .then(() => res.json({ deletedReviewId: req.params.reviewId }))
+router.delete('/:reviewId', authMiddleware, async (req, res) => {
+    const userReview = await db.Review.findById(req.params.reviewId)
+    if (userReview.senderId == req.user.id) {
+        const deletedReview = await db.Review.findByIdAndDelete(req.params.reviewId)
+        res.send('You deleted review ' + deletedReview._id)
+    } else {
+        res.status(401).json({ message: 'Invalid user or token' });
+    }
 })
 
 /* export to server.js
